CustomMethodsCreated

Description

This module has all the methods that are used throughout the application, significant ones are the methods reponsible for:

First paragraph. Parsing the JSON data.

Printing the details of songs.

Displaying all the details from the Playlist.

Written by Shashank Singh, Naman Shrivastava, Vishal and Ziyad Shaikh

Documentation

printAllGenreNames :: [GenreTableValues] -> Int -> IO ()

Prints all the Genre values that are present in the genre Table, using the GenreTableValues datatype

printAllArtistNames :: [SongCollectionArtistTrack] -> Int -> IO ()

Prints all the Artist Name values that are retirived from the entries Table, using the SongCollectionArtistTrack datatype

printSongsByGenreNames :: [SongCollectionArtistTrack] -> Int -> IO ()

Function to print the Track Name, Artist Name and Collection Name from SongCollectionArtistTrack datatype

printSongsInPlaylist :: Connection -> [UserPlaylistValues] -> IO ()

Prints the Track Name, Artist Name and Collection Name from UserPlaylistValues datatype using the printSongsByGenreNames function

printSongsByArtistForPlaylist :: [ForPlaylistAdd] -> Int -> IO ()

Prints the Track Name, Artist Name and Collection Name from ForPlaylistAdd datatype

fetchAndParseData :: Connection -> String -> IO ()

Function to Fetch and Parse the API data, which uses the functions defined in Fetch.hs and Parse.hs file. Here, the provided URL is also validated to check if or not it is empty, or returns a 404 error, and also if or not it is Valid. If any of the problems are found, same are displayed on the console along with a raw Error Message that the haskell application throws.