Database

Description

Documentation

initialiseDB

Arguments

:: IO Connection
IO Connection return type, which returns a connection to the database

Initialise DB and create tables

getOrCreateStation

Arguments

:: Connection
Connection Argument, to get the connection to the database

-> String
String Argument, containing the station name

-> String
String Argument, containing the location

-> String
String Argument, containing the state

-> IO Station
IO Station return type.

Create or Read from Station Table

saveRecords

Arguments

:: Connection
Connection Argument, to get the connection to the database

-> Vector EachRecord
Vector EachRecord Argument, containing the weather/station details to be saved

-> IO ()
IO return type.

After downloading data from web we are saving the vector record into DB

createRecord

Arguments

:: Connection
Connection Argument, to get the connection to the database

-> EachRecord
EachRecord Argument, containing the weather/station details to be inserted in the database

-> IO ()
IO return type.

Get or Create Station
Use Primary Key of Station as Foreign Key of Weather
Insert into weather_details table

queryGetAllStations

Arguments

:: Connection
Connection Argument, to get the connection to the database

-> IO [Station]
IO [Station] return type.

Find all stations

queryRandStations

Arguments

:: Connection
Connection Argument, to get the connection to the database

-> IO [Station]
IO [Station] return type

Find 10 random stations

queryByState

Arguments

:: Connection
Connection Argument, to get the connection to the database

-> IO [Response]
IO [Response] return type.

Find weather details by State

queryTopMinMaxTemperatureStations

Arguments

:: Connection
Connection Argument, to get the connection to the database

-> IO [Response]
IO[Response] return type.

Find mininum and maximum temperature of stations

updateLocationForStation

Arguments

:: Connection
Connection Argument, to get the connection to the database

-> IO ()
IO return type.

Update Average Temperature of a station

queryByStation

Arguments

:: Connection
Connection Argument, to get the connection to the database

-> IO [Response]
IO [Response] return type.

Find weather details by Station ID

queryGetWeatherDetails

Arguments

:: Connection
Connection Argument, to get the connection to the database

-> IO [WeatherRecord]
IO [WeatherRecord] return type.

Extract all records from weather_details table