Fetch

Description

This module handles connecting to the URLs of iTunes API. Also, this corraborates the URL to be valid and invalid.

Written by Shashank Singh, Naman Shrivastava, Vishal and Ziyad Shaikh

Documentation

download :: URL -> IO ByteString

Function to set a connection with the specified URL for fetching the data via the API. Here, the URL is also being checked, if or not it is a Valid URL. If not, it send back "Invalid URL" string packed as type Data.ByteString.Lazy.Char8