HTTPRequestModule

Description

This module provides all the methods necessary for filling the database with movies and actors

Written by Liam Kelly, Johannes Hartmann

Documentation

httpGetListOfMovies :: String -> IO [Movie]

Looks up all movies starting from a given time string in format yyyy-mm-dd at the TheMovieDB API.

httpGetListOfActores :: [Movie] -> IO [Actor]

Looks up all actors playing in the given list of movies at the TheMovieDB API.

getDateString :: DateTime -> String

For a given DateTime value from the Data.DateTime library a time string in yyyy-mm-dd format will be returned