HTTPRequestModule2

Description

This module provides all the necessary methods to look up the list of cinemas that play a given movie in a given area

Written by Manuel Campos Villarreal, Johannes Hartmann

Documentation

httpGetCinemaList :: String -> IO [Cinema]

for a given location this function returns a list of cinemas that are nearby

httpApiCinemaRequest :: Movie -> [Cinema] -> IO [Cinema]

for a given movie and a list of cinemas this function returns a list of cinemas that play this movie