IOActionModule

Description

This module provides all the necessary functions for handling the user interaction

Written by Johannes Hartmann

Documentation

askForLocation :: IO String

This function asks the user for a location and returns it as a string

askForActor :: IO String

This function asks the user for an actor and returns it as a string

askToSelectMovie :: IO Int

This function asks the user to select a movie and returns it as an int

printCinemas :: [Cinema] -> IO ()

This function prints a given list of cinemas to std out

printMovies :: [Movie] -> IO ()

This function prints a given list of movies to std out