JSONParserModule

Description

This module takes care of the json parsing for each http module

Written by Liam Kelly, Manuel Campos Villarreal, Johannes Hartmann

Documentation

parseMovies :: ByteString -> [Movie]

This function parses a given byte string representing a JSON into a list of movies

parseActors :: ByteString -> Movie -> [Actor]

This function parses a given byte string representing a JSON and a movie into a list actors

parsePages :: ByteString -> Int

This function parses a given byte string representing a JSON into the total number of pages

parseCinemas :: ByteString -> [Cinema]

This function parses a given byte string representing a JSON into a list of cinemas

parseMovies2 :: ByteString -> [Movie2]

This function parses a given byte string representing a JSON into a list of movies2

Orphan instances

FromJSON Cinema

Methods

parseJSON :: Value -> Parser Cinema

FromJSON Movie2

Methods

parseJSON :: Value -> Parser Movie2

FromJSON Movie

Methods

parseJSON :: Value -> Parser Movie