Logger

Description

This module handles the output of the app. It gives the user options to choose between different output choices.

Documentation

dispalyResult :: [Message] -> IO ()

Displays total messages sent in the social network.

listHead :: String

Displays headers for the table for all the messages exchanged. Used inside dispalyResult.

resultToString :: Message -> String

Parses a Message type to a string. Used inside dispalyResult.

displayCount :: [([String], Int)] -> String -> IO ()

Displays total messages sent or received by a user in the social network.

listHeadCount :: String -> String

Displays headers for the messages sent/recieved count table. Used inside displayCount.

resultToStringCount :: ([String], Int) -> String

Parses a touple to a string. Used inside displayCount

logReceivedOutput :: [Message] -> IO ()

Logs output for Count of messages each user Received

logSentOutput :: [Message] -> IO ()

Logs output for Count of messages each user Sent

outputLogger :: [Message] -> MVar Bool -> IO ()

Handles user input