Main

Description

This purpose of the application is to download all the recent movies from the TheMovieDB API and stores them into a SQLite database. Main Functionality is as Follows: The user types in an actor in the Console; The application will look this actor up and print out all movies he's playing in. Then, for a given location it suggests cinemas in the areas that play this movie.

Written by Johannes Hartmann, Liam Kelly, Manuel Campos Villarreal

Documentation

run :: IO ()

The real application function. This function is responsible for filling up the database and interacting with the user. This function throws lots of exceptions witch will be handled in the main function

main :: IO ()

The main function of this application