Parse

Description

Documentation

parseRecords

Arguments

:: ByteString
The Data.ByteString.Lazy ByteString Argument

-> Either String (Vector EachRecord)
The return value

The parseRecords function parses the downloaded csv file into Haskell data types.

stationJsonParser

Arguments

:: Connection
The connection argument, for connecting to the database

-> IO ()
The return value

The Station function fetches the station details from database, parses the recieved details and stores them to a file.

weatherJsonParser

Arguments

:: Connection
The connection argument, for connecting to the database

-> IO ()
The return value

The weatherJsonParser function fetches the Weather details from database, parses the recieved details and stores them to a file.

writeToExistingFile

Arguments

:: [String]
The [String] argument, containing the file paths

-> ByteString
The ByteString argument, containing the json data in ByteString

-> IO ()
The return value

The writeToExisitingFile function writes the JSON data to an exisiting file in the specified directory

writeToNewFile

Arguments

:: String
The String argument, containing the name of the file

-> ByteString
The ByteString argument, containg the json data in ByteString

-> IO ()

The writeToNewFile function writes the JSON to a new file in the specified directory

removeIfExists

Arguments

:: FilePath
The file path argument

-> IO ()
The return value

The removeIfExists function removes the file if it exists in the path

Orphan instances

ToJSON Station

Instance details

Methods

toJSON :: Station -> Value

toEncoding :: Station -> Encoding

toJSONList :: [Station] -> Value

toEncodingList :: [Station] -> Encoding

ToJSON WeatherRecord

Instance details

Methods

toJSON :: WeatherRecord -> Value

toEncoding :: WeatherRecord -> Encoding

toJSONList :: [WeatherRecord] -> Value

toEncodingList :: [WeatherRecord] -> Encoding