Types

Description

Documentation

data Entry

The Entry data type is used to create rows in the entries table. Use case : Database.createRecord

Constructors

Entry

Fields

weight_ :: Maybe Int

runtime_ :: Maybe Int

premiered_ :: Maybe String

ended_ :: Maybe String

language_ :: Maybe String

site_ :: Maybe String

status_ :: Maybe String

fk_program1 :: Maybe Int

data Program

The Program data type is used to support the conversion for SQL operations.

The Program data type is used to create rows in the program table.

Use case : Database.queryProgramAllEntries, Database.queryProgramSummary

Constructors

Program

Fields

id_ :: Maybe Int

program_ :: String

data Summary

The Summary data type is used to support the conversion for SQL operations.

The Summary data type is used to create rows in the summary table.

Use case : Database.createSummary, Database.query5RandomPrograms

Constructors

Summary

Fields

summary_ :: Maybe String

fk_program2 :: Maybe Int

data Record

Used to convert raw json data into readable data type in Haskell

Use Case: Database.saveRecordsEntries, Database.saveRecordsSummary

Constructors

Record

Fields

weight :: Maybe Int

runtime :: Maybe Int

premiered :: Maybe String

ended :: Maybe String

language :: Maybe String

site :: Maybe String

status :: Maybe String

program :: String

summary :: Maybe String

data ShowWithDate

The ShowWithDate data type is used to support SQL operaions that includes dates.

Use case : Database.queryAllProgramEnded, Database.queryRunningProgram, Database.queryProgramPremieredDate

Constructors

ShowWithDate

Fields

show_id :: Maybe Int

show_title :: String

show_date :: Maybe String

data RandomSummary

The RandomSummary data type is used to support the random mini game and SQL operaions.

Use case : Database.queryProgramRandomSummary

Constructors

RandomSummary

Fields

guess_summary :: Maybe String

guess_status :: Maybe String

renameFields :: String -> String

Rename certiain fields to imporove readibility and resolve naming ambiguity

customOptions :: Options

Setting default options for FromJSON

omitNothingFields = True - Null values are now considered.