Database

Description

Documentation

initialiseDB :: IO Connection

Initialises the SQLite DB and creates table Divisions and Events if not exists

saveRecords :: Connection -> Record -> IO ()

Saves all the bank holiday data to the DB with given connection

queryBankHolidaysForCurrentYear :: Connection -> IO [Event]

Gets all the bank holidays in a division in current year

queryBankHolidaysForGivenYear :: Connection -> IO [Event]

Gets Input from user and Fetches data of all the bank holidays in a chosen division for a given year

queryNextBankHoliday :: Connection -> IO [Event]

Gets Input from user and Fetches data from DB for the next Bank Holiday

queryBankHolidays :: Connection -> String -> String -> IO [Event]

Fetches Bank holiday for the given division and year

Orphan instances

FromRow Event

Instance details

Methods

fromRow :: RowParser Event

FromRow DivisionEntry

Instance details

Methods

fromRow :: RowParser DivisionEntry

ToRow DivisionEntry

Instance details

Methods

toRow :: DivisionEntry -> [SQLData]

ToRow EventEntry

Instance details

Methods

toRow :: EventEntry -> [SQLData]