Database

Description

Database Structure Setup

Used to set up the entries, summary and program tables in SQLite.

initialiseDB :: IO Connection

Method to create database shows.db using Database.SQLite.Simple.

Database Insert

For insert of records into the newly created entries, summary and program tables.

getOrCreateProgram :: Connection -> String -> IO Program

Insert records into program table.

Able to prevent dulpication of records in the table.

createRecord :: Connection -> Record -> IO ()

Creates a record with type Entry from the Record values and inserts into the entries table.

saveRecordsEntries :: Connection -> [Record] -> IO ()

Invoke function to create entries records.

Calls upon the createRecord funtion to complete the process.

saveRecordsSummary :: Connection -> [Record] -> IO ()

Invoke function to create summary records.

Calls upon the createSummary funtion to complete the process.

createSummary :: Connection -> Record -> IO ()

Creates a record with type Summary from the Record values and inserts into the entries table.

Error handling

Allows Maybe Int to be parsed and then can check to see if the input is of type Int. If not, an error will be thrown.

maybeRead :: Read a => String -> Maybe a

Error handling for Integers

Uses Import Data.Maybe (listToMaybe) for preset of listToMaybe method.

Random Int Generator

Used during the queryProgramRandomSummary function to get the random value.

getRandomInt :: Connection -> IO Int

Generate random number between 1 and 230 (the number of records in our database).

Makes use of the listToInt function and System.Random import.

countRecords :: Connection -> IO [Program]

countRecordsInt :: Connection -> IO Int

Takes the query generated from queryRunningProgram and counts the numbers of rows produced.

Query Programs

Functions created to query the database based on the users desired functionality.

queryProgramAllEntries :: Connection -> IO [Program]

Display certain number of shows based on user input.

Includes error handeling which prevents users from entering non interger records, by using the maybeRead method.

Output is of type Program list.

queryProgramSummary :: Connection -> IO [Summary]

Show the summary of a particular show.

Includes error handeling which prevents users from entering non interger records, by using the maybeRead method.

Output is of type Summary list, to display only the show summary and record ID.

queryAllProgramEnded :: Connection -> IO [ShowWithDate]

Show all shows that have a status of Ended.

Displays output of type ShowWithDate.

queryProgramPremieredDate :: Connection -> IO [ShowWithDate]

Show all the programs premiered before user defined year.

Displays output of type ShowWithDate.

query5RandomPrograms :: Connection -> Int -> IO [Program]

Display the 4 random options + the solution record to the game

queryProgramRandomSummary :: Connection -> Int -> IO [RandomSummary]

Display the random summary and its status.

Output is a list of RandomSummary records.

countAllRunningProgram :: Connection -> IO ()

Takes the query generated from queryRunningProgram and counts the numbers of rows produced.

queryRunningProgram :: Connection -> IO [ShowWithDate]

Background method for countAllRunningProgram.

Queries all programs that are running and outputs the result with type ShowWithDate.