Database

Description

Functions

initialiseDB :: IO Connection

The initialiseDB function prepares a database in the `messages.sqlite` file to be used by the program

saveMessage :: Connection -> Message -> IO ()

The saveMessage function saves a given message into the database

selectAllMessages :: Connection -> IO Int

The selectAllMessages function gets the count of all messages in the database

selectAllMessagesByUser :: Connection -> User -> IO ()

The selectAllMessagesByUser function gets the count of messages sent and recevied by a given user and displays it to the program user

selectEachMessageSent :: Connection -> String -> IO ()

The selectEachMessageSent function gets the count of each given message and displays it to the program user