Database

Description

The module handles all the queries related to extraction from the database and also few of the queries realted to validation across the application.

The validation aids the application to print errors on the console, and avoid hard crashes.

Written by Shashank Singh, Naman Shrivastava, Vishal and Ziyad Shaikh

Documentation

initialiseDB :: IO Connection

Using the connection of a database, this function creates genre, entries and userplaylist tables if they don't already exist

getOrCreateGenre :: Connection -> String -> String -> String -> IO GenreTableValues

Inserts a only UNIQUE genres in the genre table, by checking if or not it already exists

saveRecords :: Connection -> [Record] -> IO ()

Calls the INSERT for for the values from the API Call

querySearchByAvailGenre :: Connection -> String -> IO [SongCollectionArtistTrack]

Collect Collection, Artist and Track Name using the genre received

queryGetAllGenres :: Connection -> IO [GenreTableValues]

Get all the Genre from the genre table

queryPlaylistGetValues :: Connection -> IO [UserPlaylistValues]

Collect ALL columns from userplaylist from userplaylist table, using "1" = "1" condition, which is always TRUE

queryGetAllArtists :: Connection -> IO [SongCollectionArtistTrack]

Collect ALL the Collection, Artist and Track Name from entries table, using "1" = "1" condition, which is always TRUE

querySongsForPlaylist :: Connection -> Int -> IO [SongCollectionArtistTrack]

Collect the Collection, Artist and Track Name from a particular Genre Name

queryDeleteSongFromPlaylist :: Connection -> Int -> IO ()

Delete rows from the userplaylist table for the given id this handles the exceptionw here the provied id is not present i.e. was never inserted or has already been deleted

querySearchByArtistName :: Connection -> String -> IO [ForPlaylistAdd]

Collect the values fromt the entries table

queryAddSongToPlaylist :: Connection -> Int -> String -> IO ()

Insert the fk_from_entries value in the userplaylist table, which references the entires table, which has all the Songs, Collection, Artist details