Database

Documentation

dropAllTables :: Connection -> IO ()

A function that drops the mode, route, and routesection table from the database

dumpDatabase :: Connection -> IO ()

A function that dump the data from the database into a json file

initTables :: Connection -> IO ()

A function that initializes the tables in the database

insertModes :: Connection -> [Mode] -> IO ()

A function that takes the modes list and map them to the executeInsertMode function to insert them into the table, create the table if it doesn't exist

insertRoutesByMode :: Connection -> [Route] -> IO ()

A function that takes the route list and map them to the executeInsertRoute function to insert them into the tables, create the tables if it doesn't exist

queryAllMode :: Connection -> IO [String]

A function that queries to print all of the name of the modes in the database

queryAllRoutes :: Connection -> String -> IO [String]

A function that queries to print all name of the routes from a mode in the database

createDatabase :: IO Connection

A function that creates a connection with the database

printModeName :: [String] -> IO ()

A function that prints a mode's name with the first letter in uppercase

fetchStops :: String -> String -> IO (Either String SearchDestination)

A function that queries for the search function

printMatch :: Match -> IO ()

A function to do a pretty print

queryAllDisruptions :: ByteString -> IO [DisruptionDetail]

Funtion to fetch and print all the disruptions

printDisruptions :: String -> [DisruptionDetail] -> IO ()

A function that maps the disruptions to the printDisruption to be printed for each disruption of each mode

queryAllStopPoints :: Connection -> String -> IO [String]

A function that queries for all stop points from a mode in the database