Database

Description

Functions

initialiseDB :: IO Connection

The initialiseDB function prepares a database in the `movies.sqlite` file to be used by the app

getOrCreateMovie :: Connection -> String -> Int -> IO Movie

The getOrCreateMovie function ensures the unique movie title for that year exists in the database before returning it

getLastElement :: [String] -> String

The getLastElement connects an array of strings by spaces

getOrCreateActor :: Connection -> String -> IO Actor

The getOrCreateActor function ensures the unique actor name exists in the database before returning it

iterateCast :: Connection -> Movie -> [String] -> IO ()

The iterateCast function ensures every actor in a given cast is saved into the database before associating it with a movie

getOrCreateGenre :: Connection -> String -> IO Genre

The getOrCreateGenre function ensures the unique genre exists in the database before returning it

iterateGenre :: Connection -> Movie -> [String] -> IO ()

The iterateGenre function ensures every genres in a given list is saved into the database before associating it with a movie

createRecord :: Connection -> Record -> IO ()

The createRecord function saves the full record of a given movie into the database

saveRecords :: Connection -> [Record] -> IO ()

The saveRecords function saves the full record of all movies in a list of movies into the database

databaseQueries :: Connection -> IO ()

The databaseQueries function opens the menu for the user to perform queries on the database

extractActorResults :: [Actor] -> String

The extractActorResults transforms an array of Actor's into a displayable string with the names of all the actors in the array

extractGenreResults :: [Genre] -> String

The extractGenreResults transforms an array of Genre's into a displayable string with the names of all the genres in the array

queryMovie :: Connection -> IO String

The queryMovie allows the user to see details of a movie with a specific title

queryActor :: Connection -> IO String

The queryActor allows the user to see details of an actor with a specific name

queryGenre :: Connection -> IO String

The queryGenre allows the user to see details of an genre with a specific name

queryYear :: Connection -> IO String

The queryYear allows the user to see details of a specfic year