Types

Description

Types

data User

This is the data structure that will represent a User

Constructors

User

Fields

userID :: Int

The userID method returns the id by which a user is indexed

firstName :: String

The firstName method returns the first name of a user

lastName :: String

The lastName method returns the last name of a user

data Message

This is the data structure that will represent a Message entry

Constructors

Message

Fields

msgID :: Int

The msgID method returns the id of a message

msgContent :: String

The msgContent method returns the content of a message

userFrom :: Int

The userFrom method returns the user a message was sent from

userTo :: Int

The userTo method returns the user a message was sent to