Types

Description

This module has all the custom made data types used across the applciation.

FromRow and ToRow from the Database library are extensively used to modify the datatypes for interacting witht the database.

Written by Shashank Singh, Naman Shrivastava, Vishal and Ziyad Shaikh

Documentation

data Entry

Datatype for Entries Table

Constructors

Entry

Fields

date_ :: String

collectionName_ :: String

artistName_ :: String

kind_ :: String

trackName_ :: String

collectionId_ :: Int

artistId_ :: Int

trackTimeMillis_ :: Int

isStreamable_ :: Bool

fk_primaryGenreName :: Int

data SongCollectionArtistTrack

Datatype for extracting Collection, Artist and Track Name

Constructors

SongCollectionArtistTrack

Fields

collectionNameSong :: String

artistNameSong :: String

trackNameSong :: String

data GenreTableValues

Datatype for extracting details for Genre

Constructors

GenreTableValues

Fields

id_ :: Int

country_ :: String

currency_ :: String

primaryGenreName_ :: String

data Record

Datatype for the whole details being extracted from the API call's Result values

Constructors

Record

Fields

date :: String

collectionName :: String

artistName :: String

kind :: String

trackCount :: Int

trackNumber :: Int

country :: String

currency :: String

primaryGenreName :: String

artistId :: Int

collectionId :: Int

trackId :: Int

trackName :: String

trackTimeMillis :: Int

isStreamable :: Bool

extraField :: Maybe String

data Results

Datatype for extracting details from the API call

Constructors

Results

Fields

resultCount :: Int

results :: [Record]

data GenreValues

Constructors

GenreValues

Fields

genreId :: Int

countryGenreValues :: String

currencyGenreValues :: String

primaryGenreGenreValues :: String

data UserPlaylistValues

Datatype for extracting details from userplaylist table

Constructors

UserPlaylistValues

Fields

id_USERPLAYLISTID :: Int

fk_FROMENTRIES :: Int

data ForPlaylistAdd

Constructors

ForPlaylistAdd

Fields

idForList :: Int

collectionNameForList :: String

artistNameForList :: String

trackNameForList :: String