Types

Description

Types

data Actor

This is the data structure that will represent an Actor entry

Constructors

Actor

Fields

actor_id_ :: Int

The actor_id method returns the id by which an actor entry is indexed

actor_first_name_ :: String

The actor_first_name_ method returns the first name of an actor entry

actor_last_name_ :: String

The actor_last_name_ method returns the last name of an actor entry

data Cast

A Cast entry associates an actor with a movie

Constructors

Cast

Fields

fk_movie_cast :: Int

The fk_movie_cast method returns the id of the movie in this relation

fk_actor_cast :: Int

The fk_actor_cast method returns the id of the actor in this relation

data Genre

This is the data structure that will represent a Genre entry

Constructors

Genre

Fields

genre_id_ :: Int

The genre_id_ method returns the id by which a genre entry is indexed

genre_first_name_ :: String

The genre_first_name_ method returns the first name of a genre entry

genre_second_name_ :: String

The genre_second_name_ method returns the second name of a genre entry

data MovieGenre

A MovieGenre entry associates a genre with a movie

Constructors

MovieGenre

Fields

fk_movie_genre :: Int

The fk_movie_genre method returns the id of the movie in this MovieGenre relation

fk_genre_genre :: Int

The fk_genre_genre method returns the id of the genre in this MovieGenre relation

data Movie

This is the data structure that will represent a Movie entry

Constructors

Movie

Fields

movie_id_ :: Int

The movie_id_ method returns the id by which a movie entry is indexed

title_ :: String

The title_ method returns the title of a movie entry

year_ :: Int

The year_ method returns the year of a movie entry

data Record

This is the data structure that will represent a raw Record object

Constructors

Record

Fields

title :: String

The title method returns the title present in a given json record

year :: Int

The year method returns the year present in a given json record

cast :: [String]

The cast method returns the list of actors present in a given json record

genres :: [String]

The genres method returns the list of genres present in a given json record

newtype Records

Records defines an alias for an array of Record's

Constructors

Records

Fields

records :: [Record]

The records method returns the list of records